+++ cli/src/commands/git/init.rs
@@ -258,17 +258,8 @@ async fn do_init(
                 &config_env,
             )?;
             if !workspace_command.working_copy_shared_with_git() {
-                let workspace_name = workspace_command.workspace_name().to_owned();
                 let mut tx = workspace_command.start_transaction();
-                jj_lib::git::import_head(tx.repo_mut(), &workspace_name).await?;
-                if let Some(git_head_id) = tx
-                    .repo()
-                    .view()
-                    .git_head(&workspace_name)
-                    .as_normal()
-                    .cloned()
-                {
-                    let git_head_commit = tx.repo().store().get_commit_async(&git_head_id).await?;
+                if let Some(git_head_commit) = git::import_head_commit(tx.repo_mut()).await? {
                     tx.check_out(&git_head_commit)?;
                 }
                 if tx.repo().has_changes() {